refactor(tui): Split ChannelsView into smaller components (#1590)
Extract components from ChannelsView.tsx (475 lines) into separate files:

- ChannelRow.tsx (71 lines): Single channel row in the list
- ChannelHistoryView.tsx (306 lines): Message history and compose view
- Main ChannelsView.tsx reduced to 152 lines (68% reduction)

Components extracted to src/components/channels/:
- ChannelRow: Selection indicator, unread badges, member count
- ChannelHistoryView: Message display, compose mode, mention autocomplete

All 1970 tests pass.
