+++ tui/src/components/channels/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Channel components - Extracted from ChannelsView.tsx (#1590)
+ */
+
+export { ChannelRow, type ChannelRowProps } from './ChannelRow';
+export { ChannelHistoryView, type ChannelHistoryViewProps } from './ChannelHistoryView';
